pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.43.2",
    "tld>=0.13",
    "trafilatura>=2.0.0",
    "validators>=0.34.0",
]
